separate graph record logic
